refactor: LOGIN 파일 및 폴더 정리
