+++ SERVER/src/LOGIN/db/MySQLManager.cpp
@@ -1,4 +1,4 @@
-#include "db/MySQLManager.h"
+#include "MySQLManager.h"
 #include "K_slog.h"
 #include <cstring>
