+++ SERVER/src/LOGIN/util/StringUtil.cpp
@@ -1,4 +1,4 @@
-#include "util/StringUtil.h"
+#include "StringUtil.h"
 
 std::vector<std::string> StringUtil::Split(const char *data, const int len)
 {
